+++ src/components/RouterView.js
@@ -25,6 +25,11 @@ export default () =>
     template: `
       <Element w="100%" height="100%"></Element>
     `,
+    state() {
+      return {
+        activeView: null,
+      }
+    },
     hooks: {
       ready() {
         handler = () => Router.navigate.apply(this)
@@ -34,10 +39,8 @@ export default () =>
       destroy() {
         window.removeEventListener('hashchange', handler, false)
       },
-    },
-    input: {
-      any(e) {
-        if (!this.navigating) this.parent.focus(e)
+      focus() {
+        this.activeView && this.activeView.focus()
       },
     },
   })