+++ src/lib/setup/input.js
@@ -17,12 +17,14 @@
 
 import { Log } from '../log.js'
 
+import symbols from '../symbols.js'
+
 export default (component, input) => {
-  component.prototype.___inputEvents = []
+  component.prototype[symbols.inputEvents] = []
   Object.keys(input).forEach((key) => {
     if (typeof input[key] !== 'function') {
       Log.warn(`${input[key]} is not a function`)
     }
-    component.prototype.___inputEvents[key] = input[key]
+    component.prototype[symbols.inputEvents][key] = input[key]
   })
 }